Refactor: 코드 리팩토링(기능 분리)
